refactor: move resume button from header to contact section
- Removed resume button from header component
- Added resume download button to contact section in sidebar
- Styled consistently with other contact items (GitHub, email, etc.)
- Uses document icon and same hover effects
- Better contextual placement in contact information
